do link line to register
