+++ nuxt.config.ts
@@ -16,6 +16,12 @@ export default defineNuxtConfig({
     plugins: [
       tailwindcss(),
     ],
+
+    server: {
+      allowedHosts: [
+        '56445663b9d2.ngrok-free.app' // 👈 เพิ่ม host ที่จะใช้ (จาก ngrok หรืออื่น ๆ)
+      ]
+    }
   },
 
   modules: ["@nuxt/icon"],
