+++ package.json
@@ -10,6 +10,7 @@
     "postinstall": "nuxt prepare"
   },
   "dependencies": {
+    "@line/liff": "^2.27.1",
     "@nuxt/icon": "^1.15.0",
     "@tailwindcss/vite": "^4.1.11",
     "nuxt": "^3.17.7",
